handle fractional microseconds for esc/str

+++ teensy/RemoteMode.h
@@ -9,11 +9,11 @@ public:
   virtual void end();
   virtual void execute();
   
-  void command_steer_and_esc(unsigned int str_us, unsigned int esc_us);
+  void command_steer_and_esc(float str_us, float esc_us);
 private:
   bool is_active = false;
   void update_pulses();
   unsigned long last_command_ms = 0;
-  unsigned int str_us = 1500;
-  unsigned int esc_us = 1500;
+  float str_us = 1500;
+  float esc_us = 1500;
 };
